hotfix(Tag): 	1. 매칭 생성 - 매칭별 태그 저장 로직 수정
issue: #229

+++ backend/src/main/java/com/patriot/fourlipsclover/match/entity/MatchTag.java
@@ -16,7 +16,7 @@ public class MatchTag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "match_tag_id", nullable = false)
-    private Integer matchTagId;
+    private Long matchTagId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "match_id", nullable = false)
